refactor: RandomMatchingJoinButton으로 이름 변경
